+++ CHANGES.rst
@@ -9,3 +9,4 @@ v2019.1 YYYY-MM-DD
 * Ensure Log is destroyed before its clients (GitLab #45)
 * Unify the TRS interfaces between the node and the TRSTransform classes, by
   using and exposing a TRSTransform instance inside the node class (GitLab #46)
+* Select polygon mode from GUI, and simplify it (GitLab #47)